Bugs fixed:
(i) Moving the box a little below so as to keep the numbers in the center of the box.
(ii) After each time when you click drop,the entry will be cleared.
(iii) Code-optimisation:Removal of useless function and direct calling of search fn.(As useless() is useless)
(iv) BIG BUG:
	Now you cannot click a button more than once and youcan't drop directly without picking.
	Also you can't access second player without accessing first.,etc..
	Check this out.
(v) When you give any alphabets or even nothing in the entry and drop,it will show a warning pop-up.
(vi) Finally,after ending the game,when you click OK,it will quit.
(vii)Inclusion of your 2 colours orange and green

Planning to do some more.....